requirements.txt
matplotlib>=3.5.1
Pillow>=9.1.0
numpy>=1.23.5
